backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.2
bs4==0.0.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
feedparser==6.0.12
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.0
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
